DRY fixes and refactoring
