+++ Scenes/Weapons/MovingProjectile.tscn
@@ -1,17 +1,17 @@
 [gd_scene load_steps=4 format=3 uid="uid://bi2a6hqjg0qhr"]
 
-[ext_resource type="Script" path="res://Scripts/Weapons/MovingProjectile.cs" id="1_v571t"]
+[ext_resource type="Script" path="res://Scripts/Weapons/MovingProjectile.cs" id="1_wgj7o"]
 
 [sub_resource type="CapsuleShape3D" id="CapsuleShape3D_5w67x"]
 
 [sub_resource type="StandardMaterial3D" id="StandardMaterial3D_tkiiu"]
 albedo_color = Color(0.0941176, 0.0941176, 0.0941176, 1)
 
-[node name="Projectile" type="AnimatableBody3D"]
+[node name="MovingProjectile" type="AnimatableBody3D"]
 collision_layer = 8
 collision_mask = 7
 sync_to_physics = false
-script = ExtResource("1_v571t")
+script = ExtResource("1_wgj7o")
 
 [node name="CollisionShape3D" type="CollisionShape3D" parent="."]
 transform = Transform3D(5.42438e-17, 2.62268e-09, 0.06, 0.06, -2.62268e-09, 6.03973e-17, 2.62268e-09, 0.06, -2.62268e-09, 0.00529672, 0, -0.00389753)
