+++ Scripts/Weapons/IProjectile.cs
@@ -1,4 +1,6 @@
 
+using Godot;
+
 public interface IProjectile
 {
     public int damage { get; set; }
